+++ src/TimeIt.Common/FileStatsdPayload.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace TimeIt.Core;
+namespace TimeIt.Common;
 
 public sealed class FileStatsdPayload
 {
