+++ src/TimeIt.Common/TimeIt.Common.csproj
@@ -1,5 +1,34 @@
 <Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+    <Description>Command execution time meter library</Description>
+    <Copyright>Tony Redondo @ 2023</Copyright>
+    <PackageTags>timeit, benchmark, performance</PackageTags>
+    <PackageId>TimeIt.Common</PackageId>
+    <PackageLicenseFile>LICENSE</PackageLicenseFile>
+    <DefineConstants>$(DefineConstants);NOCONSTANTS</DefineConstants>
+    <NoWarn>NETSDK1138,0436</NoWarn>
+    <InternalsAssemblyUseEmptyMethodBodies>false</InternalsAssemblyUseEmptyMethodBodies>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <InternalsAssemblyName Include="DatadogTestLogger" />
+    <InternalsAssemblyName Include="Datadog.testlogger" />
+  </ItemGroup>
+
   <ItemGroup>
     <PackageReference Include="CliWrap" Version="3.6.4" />
+    <PackageReference Include="MathNet.Numerics" Version="5.0.0" />
+    <PackageReference Include="Spectre.Console" Version="0.47.0" />
+    <PackageReference Include="IgnoresAccessChecksToGenerator" Version="0.6.0" />
+    <PackageReference Include="DatadogTestLogger" Version="0.0.36" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Include="..\..\LICENSE" Pack="true" PackagePath="\" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\TimeIt.StartupHook\TimeIt.StartupHook.csproj" />
   </ItemGroup>
 </Project>
