--- src/TimeIt.Core/TimeIt.Core.csproj
@@ -1,30 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-        <Description>Command execution time meter library</Description>
-        <Copyright>Tony Redondo @ 2023</Copyright>
-        <PackageTags>timeit, benchmark, performance</PackageTags>
-        <PackageId>TimeIt.Core</PackageId>
-        <PackageLicenseFile>LICENSE</PackageLicenseFile>
-        <DefineConstants>$(DefineConstants);NOCONSTANTS</DefineConstants>
-        <NoWarn>NETSDK1138</NoWarn>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="MathNet.Numerics" Version="5.0.0" />
-        <PackageReference Include="Spectre.Console" Version="0.47.0" />
-        <PackageReference Include="System.CommandLine" Version="2.0.0-beta4.22272.1" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <None Include="..\..\LICENSE" Pack="true" PackagePath="\" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <ProjectReference Include="..\TimeIt.Common\TimeIt.Common.csproj" />
-        <ProjectReference Include="..\TimeIt.StartupHook\TimeIt.StartupHook.csproj" />
-        <ProjectReference Include="..\TimeIt.DatadogExporter\TimeIt.DatadogExporter.csproj" />
-    </ItemGroup>
-
-</Project>
