--- src/TimeIt.DatadogExporter/TimeIt.DatadogExporter.csproj
@@ -1,23 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <InternalsAssemblyUseEmptyMethodBodies>false</InternalsAssemblyUseEmptyMethodBodies>
-        <NoWarn>0436</NoWarn>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <InternalsAssemblyName Include="DatadogTestLogger" />
-        <InternalsAssemblyName Include="Datadog.testlogger" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <PackageReference Include="IgnoresAccessChecksToGenerator" Version="0.6.0" />
-        <PackageReference Include="DatadogTestLogger" Version="0.0.36" />
-        <PackageReference Include="Spectre.Console" Version="0.47.0" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <ProjectReference Include="..\TimeIt.Common\TimeIt.Common.csproj" />
-    </ItemGroup>
-
-</Project>
